fix: resolve all Dart compilation errors for Flutter 3.44.2
- Hide BarcodeType from mobile_scanner import (name conflict with entities.dart)
- Wrap BarcodeFormat.all in list (formats expects List<BarcodeFormat>)
- Remove CupertinoPageTransitionsBuilder (removed in Flutter 3.44)
- Cast slope to double in DegradationForecast constructor
- Remove const from double-keyed map (Dart 3 restriction)

+++ labelqc_pro/lib/services/spc/spc_and_recommendations.dart
@@ -200,7 +200,7 @@ class SPCAnalyzer {
     }
 
     return DegradationForecast(
-      slope: slope,
+      slope: slope.toDouble(),
       isDecreasing: slope < -0.05,
       estimatedControlsToGradeD: controlsToD,
       confidence: r2.clamp(0, 1).toDouble(),
